functions of author crud move 'libraryContent' component to 'authors' component
